move testing codecs to own module (to allow easier reusage)
