+++ brewpi/connector/test/connector_test.py
@@ -1,18 +1,24 @@
+#!/usr/bin/env python3
+
 import logging
 import os
 import sys
 from time import sleep
 
-from brewpi.connector.controlbox.objects import MixinController
-from brewpi.protocol.factory import all_sniffers
 from controlbox.config.config import configure_module
 from controlbox.connector.socketconn import TCPServerEndpoint
 from controlbox.connector_facade import ControllerDiscoveryFacade
 from controlbox.controller import Controlbox, LongDecoder, ShortDecoder
-from controlbox.events import ControlboxEvents, ConnectorCodec, ConnectorEventVisitor
+from controlbox.events import ControlboxEvents, ConnectorEventVisitor
 from controlbox.protocol.controlbox import ControlboxProtocolV1
 from controlbox.protocol.io import determine_line_protocol
-from controlbox.support.mixins import StringerMixin, CommonEqualityMixin
+
+from brewpi.connector.controlbox.objects import MixinController
+from brewpi.connector.test.codecs import (
+    BlockBufferCodec, TypeMappingCodec, DictionaryMappingCodec,
+    ScaledTimeCodec, BrewpiStateCodec, BrewpiConstructorCodec
+)
+from brewpi.protocol.factory import all_sniffers
 
 logger = logging.getLogger(__name__)
 
@@ -32,64 +38,6 @@ def dump_device_info_typed_controller(connector, protocol: ControlboxProtocolV1)
     endpoint = connector.endpoint
     str_id = ''.join('{:02x}'.format(x) for x in id)
     logger.info("device at '%s' id '%s' current time %s " % (endpoint, str_id, time))
-
-
-class BlockBufferCodec(ConnectorCodec):
-    def encode(self, type, value):
-        return value
-
-    def decode(self, type, data, mask=None):
-        return data
-
-
-class TypeMappingCodec(ConnectorCodec):
-    def __init__(self, codecs: callable):
-        self.codecs = codecs
-
-    def encode(self, type, value):
-        delegate = self.codecs(type)
-        return delegate.encode(type, value)
-
-    def decode(self, type, data, mask=None):
-        delegate = self.codecs(type)
-        return delegate.decode(type, data, mask)
-
-
-class DictionaryMappingCodec(TypeMappingCodec):
-    def __init__(self, codecs: dict):
-        def lookup(type):
-            return codecs.get(type)
-        super().__init__(lookup)
-
-
-class BaseState(CommonEqualityMixin, StringerMixin):
-    pass
-
-
-class ScaledTime(BaseState):
-    def __init__(self, time, scale):
-        self.time = time
-        self.scale = scale
-
-
-class ScaledTimeCodec(ConnectorCodec):
-    def decode(self, type, data, mask=None):
-        time = LongDecoder()._decode(data[0:4])
-        scale = ShortDecoder()._decode(data[4:6])
-        return ScaledTime(time, scale)
-
-
-class BrewpiStateCodec(DictionaryMappingCodec):
-    def __init__(self):
-        super().__init__({
-            0: BlockBufferCodec(),
-            1: ScaledTimeCodec()
-        })
-
-
-class BrewpiConstructorCodec(DictionaryMappingCodec):
-    def __init__(self):
-        super().__init__(dict())
 
 
 class BrewpiEvents(ConnectorEventVisitor):
